pyproject.toml
[project]
name = "bookcast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.20.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "langgraph>=0.5.1",
    "langsmith>=0.4.4",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "pydub>=0.25.1",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "tiktoken>=0.9.0",
]

[tool.uv]
package = true
dev-dependencies = [
    "freezegun>=1.5.2",
    "ipython>=9.3.0",
    "matplotlib>=3.10.3",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.11.13",
    "watchdog>=6.0.0",
]

[tool.uv.sources]
bookcast = { workspace = true }

[tool.ruff]
lint.select = [
    "C9",
    "E",
    "F",
    "W",
    "I",
]
include = ["*.py"]
line-length = 120

[tool.ruff.format]
quote-style = "double"
